added mixin for easier integration
